// app.revolut.com/static/js/pages/SignUp/enums.ts
export enum SignUpScreen {
  Passcode = 'passcodeScreen',
  HomeAddress = 'homeAddressScreen',
  Name = 'nameScreen',
  BirthDate = 'birthDateScreen',
  Email = 'emailScreen',
  USCode = 'usCodeScreen',
  Disclosure = 'disclosureScreen',
  Occupation = 'occupationScreen',
  Purpose = 'purposeScreen',
  Success = 'successScreen',
  WaitingList = 'waitingListScreen',
}